make usage of serialport and crc optional
