+++ Cargo.toml
@@ -3,3 +3,5 @@ members = ["ublox", "ublox_derive"]
 
 [patch.'crates-io']
 ublox_derive = { path = "ublox_derive" }
+ublox = { path = "ublox" }
+
