+++ ublox/Cargo.toml
@@ -6,14 +6,14 @@ edition = "2018"
 license = "MIT"
 description = "A crate to communicate with u-blox GPS devices using the UBX protocol"
 
+[features]
+default = []
+serial = ["serialport", "crc"]
+
 [dependencies]
 serde = "1.0"
 serde_derive = "1.0"
-serialport = "3.3.0"
 bincode = "1.2.1"
 chrono = "0.4"
-crc = "1.8.1"
-syn = "1.0.14"
-ublox_derive = "0.0.0"
-num-traits = "0.2.11"
-num-derive = "0.3.0"
+serialport = { version = "3.3.0", optional = true }
+crc = { version = "1.8.1", optional = true }
